Rename CallbackUtils to HashingUtils
